// frontend/src/lib/api/model/chapterDto.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * e-proshno API
 * OpenAPI spec version: v1
 */
import type { TopicDto } from './topicDto';

export interface ChapterDto {
  id: string;
  subjectId: string;
  number: number;
  nameBn: string;
  label: string;
  isCustom: boolean;
  topics: TopicDto[];
}